requirements.txt
streamlit==0.79.0
pandas==1.2.3
numpy==1.18.5
matplotlib==3.3.2
Pillow==8.3.1
scikit_learn==0.24.1
nltk==3.6.2
